Corrected LCD rendering (was wrong, see #48)

+++ platform.c
@@ -38,7 +38,7 @@
 
 #include <math.h>
 
-float round (float v)
+double round (float v)
 {
 	return floor(v+0.5f);
 }
